typer updates with instance creation and checks

+++ ec2/key_pair.py
@@ -7,27 +7,32 @@ import botocore
 load_dotenv()
 ec2 = ec2_client()
 
-
 KEY_FILE = os.getenv('KEY_FILE')
 KEY_NAME = os.getenv('KEY_NAME')
 
 def key_pair_generation():
-    try:
-        key_pair = ec2.create_key_pair(KeyName=KEY_NAME)
-        private_key = key_pair["KeyMaterial"]
-        with open(KEY_FILE,"w") as write_file:
-            write_file.write(private_key)
+    response_key_pair_existence = check_key_pair()
+    if(response_key_pair_existence):
+        return 0
+    else:
+        print("! Key-pair doesn't exist\n")
+        try:
+            print("! Creating Key-pair")
+            key_pair = ec2.create_key_pair(KeyName=KEY_NAME)
+            private_key = key_pair["KeyMaterial"]
+            with open(KEY_FILE,"w") as write_file:
+                write_file.write(private_key)
 
-        os.chmod(KEY_FILE,stat.S_IRUSR | stat.S_IRUSR)
-        print(f" + Saved private key to file {KEY_FILE} with (chmod 600)")
-        return KEY_NAME
+            os.chmod(KEY_FILE,stat.S_IRUSR | stat.S_IRUSR)
+            print(f"+ Saved private key to file {KEY_FILE} with (chmod 600)")
+            return KEY_NAME
 
-    except botocore.exceptions.ClientError as e:
-        print(" :: Error::",e)
+        except botocore.exceptions.ClientError as e:
+            print(" :: Error::",e)
 
 def check_key_pair():
     if(os.path.exists(KEY_FILE)):
-        print("! Key-pair exists")
+        print("! Key-pair exists with the name provided in the .env!")
+        return True
     else:
-        print("+ Creating Key-pair")
-        key_pair_generation()
+        return False
